+++ backend/src/main/java/com/voiceprint/backend/domain/Repository/EmotionRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface EmotionRepository extends JpaRepository<Emotion, Long> {
+public interface EmotionRepository extends JpaRepository<Emotion, Byte> {
 
     Optional<Emotion> findByName(String name);
 }
